ai-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
transformers==4.37.2
torch==2.6.0
numpy==1.26.3
python-dotenv==1.0.0
pydantic==1.10.13
scikit-learn==1.4.0
pandas==2.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
streamlit==1.31.1
requests==2.31.0

